+++ src/utilities.js
@@ -1,0 +1,61 @@
+/*  A series of utilities and polyfills used by the main app */
+/*  TODO: Test HTML5 template element Polyfill on IE */
+/**
+* HTML5 template element Polyfill by Brian Blakely. See <https://jsfiddle.net/brianblakely/h3EmY/>
+* @param  {object}  d  The document.
+*/
+(function templatePolyfill(d) {
+    if ('content' in d.createElement('template')) {
+        return false;
+    }
+
+    var qPlates = d.getElementsByTagName('template'),
+        plateLen = qPlates.length,
+        elPlate,
+        qContent,
+        contentLen,
+        docContent;
+
+    for (var x = 0; x < plateLen; ++x) {
+        elPlate = qPlates[x];
+        qContent = elPlate.childNodes;
+        contentLen = qContent.length;
+        docContent = d.createDocumentFragment();
+
+        while (qContent[0]) {
+            docContent.appendChild(qContent[0]);
+        }
+
+        elPlate.content = docContent;
+    }
+})(document);
+/*!
+ * Deep merge two or more objects into the first.
+ * (c) 2019 Chris Ferdinandi, MIT License, https://gomakethings.com
+ * @param   {Object} objects  The objects to merge together
+ * @returns {Object}          Merged values of defaults and options
+ */
+var deepAssign = function () {
+
+    // Make sure there are objects to merge
+    var len = arguments.length;
+    if (len < 1) return;
+    if (len < 2) return arguments[0];
+
+    // Merge all objects into first
+    for (var i = 1; i < len; i++) {
+        for (var key in arguments[i]) {
+            // If it's an object, recursively merge
+            // Otherwise, push to key
+            if (Object.prototype.toString.call(arguments[i][key]) === '[object Object]') {
+                arguments[0][key] = deepAssign(arguments[0][key] || {}, arguments[i][key]);
+            } else {
+                arguments[0][key] = arguments[i][key];
+            }
+        }
+    }
+
+    return arguments[0];
+
+};
+