+++ CHANGELOG.md
@@ -3,12 +3,26 @@
 Project: Evaluador de Transparencia Activa en Colombia - Iniciativa MOTA
 Description: Evaluador de Transparencia Activa en Colombia - Iniciativa MOTA
 Version: 0.2.0
+URL: https://github.com/Dejusticia/mota-evaluador-publico
 Keywords: transparency, transparencia, transparencia activa, active transparency, monitor, veeduria, Colombia
 
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/)
 
 ## Unreleased
+
+### Added
+- Version information to HTML head
+- Javascript utilities file:
+  - HTML5 template element Polyfill
+  - Deep merge for objects by Chris Ferdinandi
+
+### Changed
+- Added rules to ruleset-general.json and dummy report;
+- Improved javascript inline documentation
+
+### Fixed
+- Rogue markdown markup on index.html
 
 ## [2019-05-23] - 0.2.0